Update evergreen module
This rehauls the Evergreen module into something simpler and more efficient.
First, all of the default REST endpoints have been wrapped in "query_*" methods,
which simply hits the endpoint, retrieves the information and transforms the
JSON response into a Python dict. It handles all of the HTTP requests and JSON
transformations transparently.

Second, most of the functions that involve the use of the endpoint data have
been reworked into generators. They return the most useful information: the name
of the build variant, task or test, along with the corresponding Evergreen ID.

A new endpoint has been added to get performance task information for the
performance project.

Renamed `evergreen.EvergreenClient` to `evergreen.Client` for simplicity.

`evergreen.Empty` is a subclass of `evergreen.EvergreenError` to be both more
concise and more descriptive.