PERF-248 mongo-perf overrides against git commit
You can now specify a Git commit to use as the new reference value instead of
just a tag. The checking at the moment is very naive: if a dash ("-") appears
in the reference, it is treated as a tagged baseline; otherwise, it is used as a
Git commit.

Only two arguments are required now on the command line: the reference value to
use and the JIRA ticket number. (You no longer need to pass in the Git commit of
the failing test.)