PERF-400 Post-refactoring cleanup
- move setting.sh back under each ./clusters/*/ dir
  - Not just to make Rui happy, but clusters/*/update_run_config.sh
    depends on it, so cleaner like this.
- Indent lines inside if clauses
- Straighten out some path dependencies between work dir and bin dir,
  in particular:

The outcome of this patch is that:

mongo/etc/*.yml files always enter the correct work dir under
dsi/clusters/, then call shell scripts with relative path:

    cd clusters/<variant>/
    ../../bin/setup-cluster.sh

The scripts in bin/ assume that the caller is in the correct work
dir (where terraform and other config files reside).

DSI scripts however, don't assume any particular relation between
work dir and bin dir paths. The pattern

    BINDIR=(dirname $0)
    $BINDIR/other-script.sh

is used to call other scripts under bin, and a simple

    ./yet-another-script.sh

is sufficient for a few scripts that reside in work dir.

Scripts in work dir never call scripts in bin dir, so there's no
dependency that way.

I believe, but have not tested, that these cleanups are sufficient
to make bin/setup-work-env.sh work again.

In general, the goal was to achieve a state where the work dir can
be in an arbitrary path in relation to bin dir and dsi repo as a whole.
There remains one exception to this goal, which was out of scope for
PERF-400, namely several config scripts under work dir expect AWS
PEM-file to reside in ../../keys/aws.pem.