PERF-540 Add unittests. Fix broken traversal of str(conf).
Add unittests in bin/tests/test_config.py.

Early adopters of the library reported that
  conf = ConfigDict('module_name')
  str(conf) or print(conf)
wouldn't work. In other words, traversing the entire dict
was broken. There were several bugs fixed in python code,
but even more it was also a question of the .yml files in
docs/config-specs/ being incorrect in various ways. (These
files are used for testing.) For example, a
${variable.reference} was referring to a key in another file,
but said key actually didn't exist, correctly causing a KeyError.

str(conf) was added to unittests too.