PERF-444 increase Commands.FindAndModifyInserts threshold to 25%.
Current results have gone below the 20% threshold, but result is still
not reproducible on another machine (David's machine). For example,
testing focused on a drop in performance on Jul 8, that is bigger than
the typical variance of this test (both before and after the drop), about
-12%. However, when reproducing manually (on David's machine), the diff
for the same 2 builds was actually an increase of +4%.

Note: MMAP tests are noisier than WT, but for now keeping the same wide
threshold for both.

Manual repro results:
https://docs.google.com/a/10gen.com/spreadsheets/d/1Ia1KBP8bnRDkQL4Vok9MyNJgORx1rf9SxHV3x8pci6Q/edit?usp=sharing