PERF-723 Add threshold overrides also for the *_dr tasks, now that I can.
Also remove old reference and ndays overrides. They were used to
reset the comparison of negative values from the point we started
reporting latencies as negative. In the mean time the history of
those values was also switched to negative values in Evergreen
database, so the overrides can be removed now.