PERF-723 initialsync: Adding huge threshold to start_* tests.
These tests have a pattern of finishing either in 13 ms or 5000+ ms.
Since we cannot override the Previous comparison, we have no choice
but to accept that the variance is huge (as a percentage) and
set such a large threshold.

The ticket added is actually PERF-772. This is because debugging
these is impossible when mongod.log is not uploaded for initialsync.