BF-2785 Add an override and widen thresholds for some wt comparisons, for v3.2 branch.
For Update.FieldAtOffset, the baseline was way above 100%, which is not a useful comparison.

For the Insert.*Vector tasks, especially LargeDocVector, it makes sense that with oplog
on performance is only 50%, because it writes 2x the amount of data compared to standalone.
(In other words, the diff approaching 50% is a good thing, as it means there's no other
bottleneck beyond writing data to disk.)