PERF-849 Add fio network test to test_control.noise_test.yml
Unfortunately the stock fio on our AMI doesn't work with ioengine=net,
so I ended up having to compile it from source. The compiled version
is renamed "netfio".

Due to limitations where sudo works, netfio is compiled in the
terraform remote-script on all deployments, even if it's not yet
used in productions.