PERF-1196 Actually restart MongoDB between all tests
We postponed restarting MongoDB between tests (other than initialsync) as it was found to
cause a discontinuity in ycsb results. This is expected and a major reason why we need the
restarts. In benchRun tests it seems there's no effect. This is because we already use
quiesceSystem() between tests (a w:all followed by fsync to disk on all nodes).

Confirmed in Jira discussion that this is a good time for storage team to allow a
discontinuity in timelines.