PERF-1116 Leave config servers outside the placement groups
As part of PERF-1116 we added config servers into the placement group,
which is possible now that config servers are of m5.* type. However,
since then 100% of our sharded cluster builds failed with
InsufficientInstanceCapacity. I tried a patch build and leaving the config
servers outside the placement group seems to help. So doing that with this commit.