PERF-1157 Setting thresholds for change_streams_latency task to 0.5
The current analysis system is not great for latencies. As we report
latencies as negative values, it means the "max" value will usually
be the one from the lowest thread level run. As these can be noisy,
it means the main result that gets compared is noisy. So setting
a high --threshold as well as --thread-threshold.