PERF-1347 Add find_perf_tag() to evergreen_client.Client (#84)
* Add find_perf_tag() to evergreen_client.Client

The "Add comparison..." widget needs to know the variant and task,
in addition to the tag, to successfully fetch the tag related
perf.json. This can be problematic for tags that point to patch builds
with only a few tasks, and you don't know which tasks to look for.

find_perf_tag() will brute force search all variants and tasks for
the given project and thus often succeeds in returning the version_id
that corresponds to the tag.