PERF-1711 import atlas_setup.py from mongodb_setup.py (#187)
* PERF-1711 import atlas_setup.py from mongodb_setup.py

...instead of running it separately as its own executable. Turns out
you needed to run mongodb_setup.py after all, to install a mongo shell
on workload_client. So it makes sense to not introduce a new executable.
(This deviates from the design spec.)

Includes various small fixes that brings this to a point that tests
run and complete successfully. For example, the connection_string to
exec_mongo_shell is now quoted, to allow for mongo URIs that will
contain ampersands.