TIG-1151 Update 3.2.21 link in baseline_config.yml.
The version of gperftools in MongoDB 3.2 (i.e. gperftools-2.2) may lead
to an invalid memory access on startup when running on m5 instances. A
new version of the MongoDB 3.2.21 release was generated from a patch
build that calls strlen() before trying to read from strings in the
__environ array. This change is already present in the version of
gperftools in MongoDB 3.4 and later (i.e. gperftools-2.5).

Closes #230.