Move DSI to public and open source repository
This "initial commit" is a copy of the code we have used and developed
in a non-public repository since 2015 or (earlier). The git history
will not become public but remains available internally for MongoDB
employees.

This repository is a fork of mongodb/dsi. Forks were already diverged
internally. For differences, see README.md.